+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Change Log
 
 ### Fixes
 
+- Fix an issue where “Replace All” could fail when the text to find is a combining character such as a diacritical mark.
 - [beta] Fix detailed behaviors of `^⇧↑` and `^⇧↓`  shortcuts.
 - [beta] Fix an issue the application crashed by performing `^⇧↑` or `^⇧↓` shortcut when RTL text is selected.